+++ minijinja-py/src/typeconv.rs
@@ -247,7 +247,7 @@ fn to_python_value_impl(py: Python<'_>, value: Value) -> PyResult<Py<PyAny>> {
         Ok(rv.into())
     } else {
         match value.kind() {
-            ValueKind::Undefined | ValueKind::None => Ok(().into_py(py)),
+            ValueKind::Undefined | ValueKind::None | ValueKind::Invalid => Ok(().into_py(py)),
             ValueKind::Bool => Ok(value.is_true().into_py(py)),
             ValueKind::Number => {
                 if let Ok(rv) = TryInto::<i64>::try_into(value.clone()) {