+++ minijinja/src/value/serialize.rs
@@ -1,18 +1,50 @@
 use std::collections::BTreeMap;
+use std::fmt;
 
 use serde::{ser, Serialize, Serializer};
 
 use crate::key::{Key, KeySerializer, StaticKey};
-use crate::utils::SerializationFailed;
 use crate::value::{
     value_map_with_capacity, Arc, MapType, Packed, StringType, Value, ValueMap, ValueRepr,
     VALUE_HANDLES, VALUE_HANDLE_MARKER,
 };
+
+#[derive(Debug)]
+pub struct InvalidValue(Arc<String>);
+
+impl std::error::Error for InvalidValue {}
+
+impl fmt::Display for InvalidValue {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl serde::ser::Error for InvalidValue {
+    fn custom<T>(msg: T) -> Self
+    where
+        T: fmt::Display,
+    {
+        InvalidValue(Arc::new(msg.to_string()))
+    }
+}
+
+/// Transforms a serializable value to a value object.
+///
+/// This neither fails nor panics.  For objects that cannot be represented
+/// the value might be represented as a half broken error object.
+pub fn transform<T: Serialize>(value: T) -> Value {
+    match value.serialize(ValueSerializer) {
+        Ok(rv) => rv,
+        Err(invalid) => ValueRepr::Invalid(invalid.0).into(),
+    }
+}
+
 pub struct ValueSerializer;
 
 impl Serializer for ValueSerializer {
     type Ok = Value;
-    type Error = SerializationFailed;
+    type Error = InvalidValue;
 
     type SerializeSeq = SerializeSeq;
     type SerializeTuple = SerializeTuple;
@@ -22,86 +54,86 @@ impl Serializer for ValueSerializer {
     type SerializeStruct = SerializeStruct;
     type SerializeStructVariant = SerializeStructVariant;
 
-    fn serialize_bool(self, v: bool) -> Result<Value, SerializationFailed> {
+    fn serialize_bool(self, v: bool) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::Bool(v).into())
     }
 
-    fn serialize_i8(self, v: i8) -> Result<Value, SerializationFailed> {
+    fn serialize_i8(self, v: i8) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::I64(v as i64).into())
     }
 
-    fn serialize_i16(self, v: i16) -> Result<Value, SerializationFailed> {
+    fn serialize_i16(self, v: i16) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::I64(v as i64).into())
     }
 
-    fn serialize_i32(self, v: i32) -> Result<Value, SerializationFailed> {
+    fn serialize_i32(self, v: i32) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::I64(v as i64).into())
     }
 
-    fn serialize_i64(self, v: i64) -> Result<Value, SerializationFailed> {
+    fn serialize_i64(self, v: i64) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::I64(v).into())
     }
 
-    fn serialize_i128(self, v: i128) -> Result<Value, SerializationFailed> {
+    fn serialize_i128(self, v: i128) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::I128(Packed(v)).into())
     }
 
-    fn serialize_u8(self, v: u8) -> Result<Value, SerializationFailed> {
+    fn serialize_u8(self, v: u8) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::U64(v as u64).into())
     }
 
-    fn serialize_u16(self, v: u16) -> Result<Value, SerializationFailed> {
+    fn serialize_u16(self, v: u16) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::U64(v as u64).into())
     }
 
-    fn serialize_u32(self, v: u32) -> Result<Value, SerializationFailed> {
+    fn serialize_u32(self, v: u32) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::U64(v as u64).into())
     }
 
-    fn serialize_u64(self, v: u64) -> Result<Value, SerializationFailed> {
+    fn serialize_u64(self, v: u64) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::U64(v).into())
     }
 
-    fn serialize_u128(self, v: u128) -> Result<Value, SerializationFailed> {
+    fn serialize_u128(self, v: u128) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::U128(Packed(v)).into())
     }
 
-    fn serialize_f32(self, v: f32) -> Result<Value, SerializationFailed> {
+    fn serialize_f32(self, v: f32) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::F64(v as f64).into())
     }
 
-    fn serialize_f64(self, v: f64) -> Result<Value, SerializationFailed> {
+    fn serialize_f64(self, v: f64) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::F64(v).into())
     }
 
-    fn serialize_char(self, v: char) -> Result<Value, SerializationFailed> {
+    fn serialize_char(self, v: char) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::Char(v).into())
     }
 
-    fn serialize_str(self, value: &str) -> Result<Value, SerializationFailed> {
+    fn serialize_str(self, value: &str) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::String(Arc::new(value.to_owned()), StringType::Normal).into())
     }
 
-    fn serialize_bytes(self, value: &[u8]) -> Result<Value, SerializationFailed> {
+    fn serialize_bytes(self, value: &[u8]) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::Bytes(Arc::new(value.to_owned())).into())
     }
 
-    fn serialize_none(self) -> Result<Value, SerializationFailed> {
+    fn serialize_none(self) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::None.into())
     }
 
-    fn serialize_some<T: ?Sized>(self, value: &T) -> Result<Value, SerializationFailed>
+    fn serialize_some<T: ?Sized>(self, value: &T) -> Result<Value, InvalidValue>
     where
         T: Serialize,
     {
-        value.serialize(self)
+        Ok(transform(value))
     }
 
-    fn serialize_unit(self) -> Result<Value, SerializationFailed> {
+    fn serialize_unit(self) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::None.into())
     }
 
-    fn serialize_unit_struct(self, _name: &'static str) -> Result<Value, SerializationFailed> {
+    fn serialize_unit_struct(self, _name: &'static str) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::None.into())
     }
 
@@ -110,7 +142,7 @@ impl Serializer for ValueSerializer {
         name: &'static str,
         variant_index: u32,
         variant: &'static str,
-    ) -> Result<Value, SerializationFailed> {
+    ) -> Result<Value, InvalidValue> {
         if name == VALUE_HANDLE_MARKER && variant == VALUE_HANDLE_MARKER {
             Ok(VALUE_HANDLES.with(|handles| {
                 let mut handles = handles.borrow_mut();
@@ -127,11 +159,11 @@ impl Serializer for ValueSerializer {
         self,
         _name: &'static str,
         value: &T,
-    ) -> Result<Value, SerializationFailed>
+    ) -> Result<Value, InvalidValue>
     where
         T: Serialize,
     {
-        value.serialize(self)
+        Ok(transform(value))
     }
 
     fn serialize_newtype_variant<T: ?Sized>(
@@ -140,22 +172,22 @@ impl Serializer for ValueSerializer {
         _variant_index: u32,
         variant: &'static str,
         value: &T,
-    ) -> Result<Value, SerializationFailed>
+    ) -> Result<Value, InvalidValue>
     where
         T: Serialize,
     {
         let mut map = value_map_with_capacity(1);
-        map.insert(Key::Str(variant), ok!(value.serialize(self)));
+        map.insert(Key::Str(variant), transform(value));
         Ok(ValueRepr::Map(Arc::new(map), MapType::Normal).into())
     }
 
-    fn serialize_seq(self, len: Option<usize>) -> Result<Self::SerializeSeq, SerializationFailed> {
+    fn serialize_seq(self, len: Option<usize>) -> Result<Self::SerializeSeq, InvalidValue> {
         Ok(SerializeSeq {
             elements: Vec::with_capacity(len.unwrap_or(0)),
         })
     }
 
-    fn serialize_tuple(self, len: usize) -> Result<Self::SerializeTuple, SerializationFailed> {
+    fn serialize_tuple(self, len: usize) -> Result<Self::SerializeTuple, InvalidValue> {
         Ok(SerializeTuple {
             elements: Vec::with_capacity(len),
         })
@@ -165,7 +197,7 @@ impl Serializer for ValueSerializer {
         self,
         _name: &'static str,
         len: usize,
-    ) -> Result<Self::SerializeTupleStruct, SerializationFailed> {
+    ) -> Result<Self::SerializeTupleStruct, InvalidValue> {
         Ok(SerializeTupleStruct {
             fields: Vec::with_capacity(len),
         })
@@ -177,14 +209,14 @@ impl Serializer for ValueSerializer {
         _variant_index: u32,
         variant: &'static str,
         len: usize,
-    ) -> Result<Self::SerializeTupleVariant, SerializationFailed> {
+    ) -> Result<Self::SerializeTupleVariant, InvalidValue> {
         Ok(SerializeTupleVariant {
             name: variant,
             fields: Vec::with_capacity(len),
         })
     }
 
-    fn serialize_map(self, len: Option<usize>) -> Result<Self::SerializeMap, SerializationFailed> {
+    fn serialize_map(self, len: Option<usize>) -> Result<Self::SerializeMap, InvalidValue> {
         Ok(SerializeMap {
             entries: value_map_with_capacity(len.unwrap_or(0)),
             key: None,
@@ -195,7 +227,7 @@ impl Serializer for ValueSerializer {
         self,
         _name: &'static str,
         len: usize,
-    ) -> Result<Self::SerializeStruct, SerializationFailed> {
+    ) -> Result<Self::SerializeStruct, InvalidValue> {
         Ok(SerializeStruct {
             fields: value_map_with_capacity(len),
         })
@@ -207,7 +239,7 @@ impl Serializer for ValueSerializer {
         _variant_index: u32,
         variant: &'static str,
         len: usize,
-    ) -> Result<Self::SerializeStructVariant, SerializationFailed> {
+    ) -> Result<Self::SerializeStructVariant, InvalidValue> {
         Ok(SerializeStructVariant {
             variant,
             map: value_map_with_capacity(len),
@@ -221,18 +253,17 @@ pub struct SerializeSeq {
 
 impl ser::SerializeSeq for SerializeSeq {
     type Ok = Value;
-    type Error = SerializationFailed;
+    type Error = InvalidValue;
 
-    fn serialize_element<T: ?Sized>(&mut self, value: &T) -> Result<(), SerializationFailed>
+    fn serialize_element<T: ?Sized>(&mut self, value: &T) -> Result<(), InvalidValue>
     where
         T: Serialize,
     {
-        let value = ok!(value.serialize(ValueSerializer));
-        self.elements.push(value);
+        self.elements.push(transform(value));
         Ok(())
     }
 
-    fn end(self) -> Result<Value, SerializationFailed> {
+    fn end(self) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::Seq(Arc::new(self.elements)).into())
     }
 }
@@ -243,18 +274,17 @@ pub struct SerializeTuple {
 
 impl ser::SerializeTuple for SerializeTuple {
     type Ok = Value;
-    type Error = SerializationFailed;
+    type Error = InvalidValue;
 
-    fn serialize_element<T: ?Sized>(&mut self, value: &T) -> Result<(), SerializationFailed>
+    fn serialize_element<T: ?Sized>(&mut self, value: &T) -> Result<(), InvalidValue>
     where
         T: Serialize,
     {
-        let value = ok!(value.serialize(ValueSerializer));
-        self.elements.push(value);
+        self.elements.push(transform(value));
         Ok(())
     }
 
-    fn end(self) -> Result<Value, SerializationFailed> {
+    fn end(self) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::Seq(Arc::new(self.elements)).into())
     }
 }
@@ -265,18 +295,17 @@ pub struct SerializeTupleStruct {
 
 impl ser::SerializeTupleStruct for SerializeTupleStruct {
     type Ok = Value;
-    type Error = SerializationFailed;
+    type Error = InvalidValue;
 
-    fn serialize_field<T: ?Sized>(&mut self, value: &T) -> Result<(), SerializationFailed>
+    fn serialize_field<T: ?Sized>(&mut self, value: &T) -> Result<(), InvalidValue>
     where
         T: Serialize,
     {
-        let value = ok!(value.serialize(ValueSerializer));
-        self.fields.push(value);
+        self.fields.push(transform(value));
         Ok(())
     }
 
-    fn end(self) -> Result<Value, SerializationFailed> {
+    fn end(self) -> Result<Value, InvalidValue> {
         Ok(Value(ValueRepr::Seq(Arc::new(self.fields))))
     }
 }
@@ -288,18 +317,17 @@ pub struct SerializeTupleVariant {
 
 impl ser::SerializeTupleVariant for SerializeTupleVariant {
     type Ok = Value;
-    type Error = SerializationFailed;
+    type Error = InvalidValue;
 
-    fn serialize_field<T: ?Sized>(&mut self, value: &T) -> Result<(), SerializationFailed>
+    fn serialize_field<T: ?Sized>(&mut self, value: &T) -> Result<(), InvalidValue>
     where
         T: Serialize,
     {
-        let value = ok!(value.serialize(ValueSerializer));
-        self.fields.push(value);
+        self.fields.push(transform(value));
         Ok(())
     }
 
-    fn end(self) -> Result<Value, SerializationFailed> {
+    fn end(self) -> Result<Value, InvalidValue> {
         let mut map = value_map_with_capacity(1);
         map.insert(
             Key::Str(self.name),
@@ -316,31 +344,30 @@ pub struct SerializeMap {
 
 impl ser::SerializeMap for SerializeMap {
     type Ok = Value;
-    type Error = SerializationFailed;
+    type Error = InvalidValue;
 
-    fn serialize_key<T: ?Sized>(&mut self, key: &T) -> Result<(), SerializationFailed>
+    fn serialize_key<T: ?Sized>(&mut self, key: &T) -> Result<(), InvalidValue>
     where
         T: Serialize,
     {
-        let key = ok!(key.serialize(KeySerializer));
-        self.key = Some(key);
+        match key.serialize(KeySerializer) {
+            Ok(key) => self.key = Some(key),
+            Err(_) => self.key = None,
+        }
         Ok(())
     }
 
-    fn serialize_value<T: ?Sized>(&mut self, value: &T) -> Result<(), SerializationFailed>
+    fn serialize_value<T: ?Sized>(&mut self, value: &T) -> Result<(), InvalidValue>
     where
         T: Serialize,
     {
-        let key = self
-            .key
-            .take()
-            .expect("serialize_value called before serialize_key");
-        let value = ok!(value.serialize(ValueSerializer));
-        self.entries.insert(key, value);
+        if let Some(key) = self.key.take() {
+            self.entries.insert(key, transform(value));
+        }
         Ok(())
     }
 
-    fn end(self) -> Result<Value, SerializationFailed> {
+    fn end(self) -> Result<Value, InvalidValue> {
         Ok(Value(ValueRepr::Map(
             Arc::new(self.entries),
             MapType::Normal,
@@ -351,14 +378,14 @@ impl ser::SerializeMap for SerializeMap {
         &mut self,
         key: &K,
         value: &V,
-    ) -> Result<(), SerializationFailed>
+    ) -> Result<(), InvalidValue>
     where
         K: Serialize,
         V: Serialize,
     {
-        let key = ok!(key.serialize(KeySerializer));
-        let value = ok!(value.serialize(ValueSerializer));
-        self.entries.insert(key, value);
+        if let Ok(key) = key.serialize(KeySerializer) {
+            self.entries.insert(key, transform(value));
+        }
         Ok(())
     }
 }
@@ -369,22 +396,21 @@ pub struct SerializeStruct {
 
 impl ser::SerializeStruct for SerializeStruct {
     type Ok = Value;
-    type Error = SerializationFailed;
+    type Error = InvalidValue;
 
     fn serialize_field<T: ?Sized>(
         &mut self,
         key: &'static str,
         value: &T,
-    ) -> Result<(), SerializationFailed>
+    ) -> Result<(), InvalidValue>
     where
         T: Serialize,
     {
-        let value = ok!(value.serialize(ValueSerializer));
-        self.fields.insert(Key::Str(key), value);
+        self.fields.insert(Key::Str(key), transform(value));
         Ok(())
     }
 
-    fn end(self) -> Result<Value, SerializationFailed> {
+    fn end(self) -> Result<Value, InvalidValue> {
         Ok(ValueRepr::Map(Arc::new(self.fields), MapType::Normal).into())
     }
 }
@@ -396,22 +422,21 @@ pub struct SerializeStructVariant {
 
 impl ser::SerializeStructVariant for SerializeStructVariant {
     type Ok = Value;
-    type Error = SerializationFailed;
+    type Error = InvalidValue;
 
     fn serialize_field<T: ?Sized>(
         &mut self,
         key: &'static str,
         value: &T,
-    ) -> Result<(), SerializationFailed>
+    ) -> Result<(), InvalidValue>
     where
         T: Serialize,
     {
-        let value = ok!(value.serialize(ValueSerializer));
-        self.map.insert(Key::Str(key), value);
+        self.map.insert(Key::Str(key), transform(value));
         Ok(())
     }
 
-    fn end(self) -> Result<Value, SerializationFailed> {
+    fn end(self) -> Result<Value, InvalidValue> {
         let mut rv = BTreeMap::new();
         rv.insert(
             self.variant,